+++ UnitTests/SessionStateTest.cs
@@ -159,7 +159,7 @@ namespace UnitTests
             Hashtable getTable = new Hashtable(1000);//only used in 1 thread at a time
 
             //Synchronously populate 1000 messages
-            for (int i = 1; i < 1000; i++) {
+            for (ulong i = 1; i < 1000; i++) {
                 string msg = "msg" + i;
                 state.Set(i, msg);
                 setTable[i] = msg;
@@ -170,7 +170,7 @@ namespace UnitTests
             ThreadPool.QueueUserWorkItem(delegate(object stateObject) {
                 AutoResetEvent internalSetEvent = (AutoResetEvent)((object[])stateObject)[0];
                 SessionState internalState = (SessionState)((object[])stateObject)[1];
-                for (int i = 1001; i < 2000; i++) {
+                for (ulong i = 1001; i < 2000; i++) {
                     try {
                         internalState.Set(i, "msg" + i);
                     }
@@ -188,7 +188,7 @@ namespace UnitTests
             ThreadPool.QueueUserWorkItem(delegate(object stateObject){
                 AutoResetEvent internalGetEvent = (AutoResetEvent)((object[])stateObject)[0];
                 SessionState internalState = (SessionState)((object[])stateObject)[1];
-                for (int i = 1; i < 1000; i++) {
+                for (ulong i = 1; i < 1000; i++) {
                     try {
                         List<string> lst = new List<string>(1);
                         internalState.Get(i, i, lst);
